// source/src/shared/lib/enums/LevelsRu.ts
export enum LevelsRu {
    INT = 'Простые числа',
    FRACTION = 'Дроби',
    DEGREE = 'Степени',
    SQRT = 'Квадратные корни',
    MODULE = 'Модули',
    LOGARITHM = 'Логарифмы',
    DECIMIMAL_LOGARITHM = 'Десятичные логарифмы',
    NATURAL_LOGARITHM = 'Натуральные Логарифмы',
    TRIGONOMETRIC = 'Тригонометрические функции'
}