// source/src/shared/lib/enums/Levels.ts
export enum Levels {
    INT = 'INT',
    FRACTION = 'FRACTION', // READY
    DEGREE = 'DEGREE',
    SQRT = 'SQRT', // READY
    MODULE = 'MODULE',
    LOGARITHM = 'LOGARITHM', // READY
    DECIMIMAL_LOGARITHM = 'DECIMIMAL_LOGARITHM',
    NATURAL_LOGARITHM = 'NATURAL_LOGARITHM',
    TRIGONOMETRIC = 'TRIGONOMETRIC'
}